Mi parte terminada

+++ Datos/Infrastructure/Actividad.cs
@@ -23,7 +23,7 @@ namespace Datos.Infrastructure
         public int Id_Actividad { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public Nullable<byte> Media_Valoracion { get; set; }
+        public Nullable<double> Media_Valoracion { get; set; }
         public Nullable<System.DateTime> Fecha { get; set; }
         public string DNI_Monitor { get; set; }
